divided bot.py by logic
