+++ bot_init.py
@@ -1,0 +1,6 @@
+from aiogram import Bot
+from aiogram.dispatcher import Dispatcher
+from config.config import TOKEN
+
+bot = Bot(TOKEN)
+dp = Dispatcher(bot)
