+++ messages.py
@@ -1,0 +1,9 @@
+help_message = 'Список доступных команд:\n' \
+            '/weather - прогноз погоды в локации на сегодня\n' \
+            '/rate - курс доллара на текущий момент'
+welcome_message = 'Привет {}!\n' + help_message
+
+MESSAGES = {
+    'welcome': welcome_message,
+    'help': help_message,
+}
